third_party/blink/web_tests/platform/generic/fast/css/unicode-bidi-computed-value-expected.txt

This tests the computed value of unicode-bidi property.

PASS styleOf("span", {}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: embed;"}).unicodeBidi is "embed"
PASS styleOf("span", {"style":"unicode-bidi: isolate;"}).unicodeBidi is "isolate"
PASS styleOf("span", {"style":"unicode-bidi: bidi-override;"}).unicodeBidi is "bidi-override"
PASS styleOf("span", {"style":"unicode-bidi: plaintext;"}).unicodeBidi is "plaintext"
PASS styleOf("span", {"style":"unicode-bidi: bad-value;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: embed embed;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: embed plain-text;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: bidi-override isolate;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: isolate bidi-override;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: isolate-override;"}).unicodeBidi is "isolate-override"
PASS styleOf("span", {"style":"unicode-bidi: bidi-override isolate bidi-override;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: bidi-override isolate isolate;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: bidi-override bad-value;"}).unicodeBidi is "normal"
PASS styleOf("span", {"style":"unicode-bidi: bidi-override embed;"}).unicodeBidi is "normal"
PASS successfullyParsed is true

TEST COMPLETE
